The rest of the changes to the scripts. This time compiled to be sure that nothing is missing.
beside the missing scripts and some minor fixes.

Add ForceOpenMouth property to the sslActorAlias to allow mods like DDi keep the mouth open when the actor is using some Gag and the sexla have the controll of the actor mouth.

Fix the NiOverride High Heels issue that don't refresh the Heels offset when the actor get striped from the Heels after the LeadIn

Fix the issue with the animation don't refreshing the first stage when you come back from other stage of the same animation.

Fix the issue with animations with AnimObjects type Furniture that keep the AnimObject when the animation is forced to end before the last stage.

Separate the exit idles of creatures by RaceKey to allow it being used to Reset the Default idle each time tha is needed without the issues of use the wrong idle on the wrong creature at the wrong moment.

** Some Variables have double check the null/none becaous on LE one way don't seems to be enog