Create SexlabOffsets.json
I never remember to include this file mostly because sexlab will make his own once installed with the default values given by the animators but this file have My BedOffsets for the animations i know so avoid many work for the users and looks good. (also include my furnitureOffsets that are not relevant for the framework and can be removed but is not necesary)

By the way SL Animation Load" don't ïnclude a way to add the bedoffset so just the default animations come with it.