Update the Expression System
Some fix to Enjoyment and Pain functions issues. Probably need more work.
Some fix for the Separate Orgasm. Probably need more work too.
Moved the "OpenMouth Size" option from the "Animation Settings" page to the "Expression Editor" page.
Enhance the "Expression Editor" page. Now the "Test Expression" options also allow test the expression with the Mouth Open.
Add Advanced OpenMouth configuration with separate Phonemes array for Male and Female on the "Expression Editor" page.
Add some error treatment  for the functions

WARNING
add the global to the ToFloatArray() and ToIntArray() functions on the sslBaseExpression is required and is agains the rules but shold not be a problem since those functions are not used by any mod.