Fix the Mouth issues
After many test with different configurations and with & without the MFG the conclusion about the fish mouth issue is that was caused by the ApplyPresetFloats function on each RefreshExpression, and not the MoveLips function like first thug.
Digging more become obvious that something about the functions to set the Phoneme was causing it.

The only solution that seems to be working is take the last bigger valued Phoneme and set it as last on the loop. I apply the same solution to the MoveLips function that on some expressions phases show another issue apparently related where instead of do the expected lips movement whit the settled wait times and strength, was fully opening and closing the mouth instantly.
Now everything is working but need to be tested on SE to be sure. also some values adjustments need to be done on the MoveLips function eventually but is enough for now

*** The fish mouth issue is when the human actor open and close the mouth too often like a fish trying to breath out of the water.