Adjusting Moaning and allowing orgasm at the end for actors without orgasms & enjoyment
The lips movement of the moaning was not enough for some users o mod scenes like those of the pregnancy mod.

Allowing the orgasm at the end with the "separate orgasm" option enabled for actors that are allowed for orgasm but don't reach enough enjoyment to have at least one orgasm before the last stage.