Update sslActorAlias.psc
Fix illumination issue on the actors after teleport.
Update advanced Lips Sync options effects on the actors

Adjustments on the enjoyment values. By the way SexLab isn't made to deal with the "Menu Mode" because the Real Time Clock keep running OnMenuMode like the commented log show and that cause issues on the animations stages and the enjoyment between  other thing.