Update sslBaseVoice.psc
Add few new Extended functions for mod authors that want more control over the moaning effects.

Fix compatibility issue with the deprecated Moan function.  Now the Moan function will use the values of the version 1.62 again because some of the mods using the function are made with those values in main.