Fix Stats problems and Enhance the MCM
Fix the problem with the "TimeSpent" stats becoming negative after sex.

Allow edit some of the actor stats while SexLab is in Debug Mode. Is very basic and don't allow 0 values for tech reasons but is probably better this way. (Also probably need the real SkyUI SDK only to be compiled because I think the public one don't come the Input options)

Fix some errors on the SexLab MCM and add new enhancements .