Fix problem with the RealTime native function on AE
Users on AE are reporting issues with the Orgasm and the automatic transition between animations stages. Ed86 do few test and seems like the GetCurrentGameRealTime function is not working on AE.

I set again my version of the function that is working fine on SE, that way you can use it as reference if the Native is not working for real.

Remember I can't test AE right now, so I can't know the real reason!